# app/Virtual/Resources/BookResource.php
<?php
/**
 * @OA\Schema(
 *     title="BookResource",
 *     description="Book resource",
 *     @OA\Xml(
 *         name="BookResource"
 *     )
 * )
 */
class BookResource
{
    /**
     * @OA\Property(
     *     title="Data",
     *     description="Data wrapper"
     * )
     *
     * @var \App\Virtual\Models\Book[]
     */
    private $data;
}